src/requirements.txt
apache-beam==2.45.0
asn1crypto==1.5.1
azure-devops==6.0.0b4
azuredevopsX==0.0.17
beam-nuggets==0.18.1
certifi==2022.12.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
factory-boy==3.2.1
Faker==17.6.0
fastavro==1.7.2
fasteners==0.18
greenlet==2.0.2
grpcio==1.51.3
hdfs==2.7.0
httplib2==0.20.4
idna==3.4
isodate==0.6.1
kafka-python==2.0.2
msrest==0.6.21
numpy==1.22.4
oauthlib==3.2.2
objsize==0.6.1
orjson==3.8.7
pg8000==1.29.4
proto-plus==1.22.2
protobuf==3.20.3
pyarrow==9.0.0
pydot==1.4.2
pymongo==3.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requestx==0.0.9
scramp==1.4.4
six==1.16.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.40.0
typing_extensions==4.5.0
urllib3==1.26.14
zstandard==0.20.0
